refactor-1, moved components out
